More partially built sandbox
Every damn CVariable related edit is bad, as usual with their containerized variable system. Just not built properly.....

Disabled vector set-limits in a few places.

Hackity-hack-hack stuff all over the place in Variable.h just to to compile.....AAAAAAAARGH
All cuz of 'limits' not using the variable type as a base......ugh......

'Sandbox' project built at last. Now for all dem plugins.

+++ Code/Sandbox/EditorQt/Objects/EntityObject.cpp
@@ -5972,7 +5972,7 @@ string CEntityObject::GetMouseOverStatisticsText() const
 					IGeomCache::SStatistics stats = pGeomCache->GetStatistics();
 
 					string averageDataRate;
-					FormatFloatForUI(averageDataRate, 2, stats.m_averageAnimationDataRate);
+					FormatFloatForUI(averageDataRate, 2, (double)stats.m_averageAnimationDataRate);
 
 					statsText += "\n  " + (stats.m_bPlaybackFromMemory ? string("Playback from Memory") : string("Playback from Disk"));
 					statsText += "\n  " + averageDataRate + " MiB/s Average Animation Data Rate";
